fix: SonarQube fixes as follow up to #3337 (#3339)
This has a number of small follow-up fixes after https://github.com/acts-project/acts/pull/3337 that I did not
want to include in that PR.

Blocked by:
- #3337

+++ Core/include/Acts/Utilities/BinningData.hpp
@@ -269,7 +269,7 @@ class BinningData {
       return (eta(position));
     }
     if (toUnderlying(binvalue) < 3) {
-      return position[toUnderlying(binvalue)];
+      return static_cast<float>(position[toUnderlying(binvalue)]);
     }
     // phi gauging
     return phi(position);
